Added functionality to accept stdin for the json configuration
